feat: Add on boarding module
This commit introduces the on boarding feature module to the project.
The commit creates a new module, `on_boarding`, and adds the necessary build configurations, source files, and test files for a standard Android library module.
It also includes necessary project files such as gitignore, the Android manifest, consumer-rules, and proguard-rules.
Additionally, the module has been added to the settings.gradle.kts to let Gradle know it is a part of the project.