feat: Update Remote Appointment Data Source to return Flow of Resource
This commit modifies the `RemoteAppointmentsDataSource` interface to return a `Flow` of `Resource` wrapping a list of `Appointment` objects, instead of just returning a `Flow` of a list of `Appointment` objects. This change applies to the `getPatientAppointments`, `getAppointments`, and `getAppointments` with statusId functions, enhancing the ability to handle loading and error states from the remote data source.