feat: Update register parameters in AuthRepository
This commit updates the `register` function in `AuthRepository` and its implementation. The order of the parameters has been modified, and additional parameters have been added: `nationalId`, `phone`, and `address`. The `password` parameter is now moved to the end of the parameters list. The implementation in `AuthRepositoryImpl` has been updated accordingly.