feat: Add time slot division functionality
This commit introduces the ability to divide a `TimeSlot` into smaller time slots of a specified duration. A new `divide` extension function has been added to the `TimeSlot` model to facilitate this. It takes the desired duration of each smaller slot in minutes and returns a list of new `TimeSlot` objects. In addition, a new function has been added `getFormattedTime` to format time.