feat: Add PharmacyViewModel
This commit introduces the `PharmacyViewModel` class, which handles the logic for managing pharmacy data. It includes functionality for adding, updating, and listing pharmacies. The ViewModel uses use cases for interacting with pharmacy data and exposes state flows to reflect the results of these operations. It contains functions `addNewPharmacy`, `updatePharmacy` and `listPharmacies`. It also has the states of pharmacy properties `idState`, `nameState`, `phoneState`, `addressState`, `longitudeState`, `latitudeState` and `contractStatusIdState`.