feat: Add Price to Doctor and DoctorEntity
This commit introduces a `price` field to the `Doctor` class in the domain layer and the `DoctorEntity` in the data layer. This allows for storing and retrieving the price associated with a doctor. The `price` field is added to the data layer and its default value is `0.0`. Also, the `price` attribute was added in the DoctorEntity to Doctor mapper and vice versa.