refactor: Rename LoginInfoResponse to UserResponse and Update API Service
This commit renames `LoginInfoResponse` to `UserResponse` and modifies the `ApiService` to use the new `UserResponse` type.
It updates the `getLoginInfo` function in `ApiService` to return `UserResponse` instead of `LoginInfoResponse`.
It also renames `LoginInfoResponse` file to `UserResponse` and `LoginInfoResponse` class to `UserResponse` class.
The `UserResponse` data class now includes a `User` field.