refactor: Use LaunchedEffect with Resource.Success for Loading State
This commit modifies the `DoctorsContent` composable to use `LaunchedEffect` with `actionResource` as the key when the resource state is `Resource.Success`. This ensures that the loading state is updated correctly upon successful completion of an action. Previously, `LaunchedEffect(key1 = Unit)` was used, which didn't tie the effect to the specific resource update. This change provides a more precise way of managing the loading state based on the success of a resource operation.