feat: Introduce generic Result sealed interface
This commit introduces a new sealed interface `Result` in the `core.domain.model` package.

`Result` is a generic type that can represent either a `Success` with data of type `D` or an `Error` with an error of type `E`.

It also includes the following extension functions:
- `map`: Transforms the data of a `Success` result.
- `asEmptyDataResult`: Converts a `Result` to an `EmptyResult` (Result<Unit, E>).
- `onSuccess`: Performs an action if the result is `Success`.
- `onError`: Performs an action if the result is `Error`.

A typealias `EmptyResult<E>` is defined for `Result<Unit, E>`.