refactor: Rename SignInAction navigation events
This commit refactors the `SignInAction` sealed interface to improve the clarity of navigation-related actions.

Specifically, the following changes were made:
- `SignInAction.ResetPassword` has been renamed to `SignInAction.NavigateToPasswordRecoveryScreen`.
- `SignInAction.CreateAccount` has been renamed to `SignInAction.NavigateToSignUpScreen`.

These changes are reflected in `SignInScreen.kt` where these actions are handled and in `SignInViewModel.kt` within the `onAction` method. The renaming provides a more descriptive representation of the actions' purposes, indicating they are responsible for triggering navigation events.