feat: Add JaCoCo convention plugin for Android applications
This commit introduces a new Gradle convention plugin, `AndroidApplicationJacocoConventionPlugin.kt`, to configure JaCoCo for Android application modules.

The plugin applies the `jacoco` plugin and configures the `ApplicationExtension` to enable code coverage for both Android tests and unit tests in all build types. It then utilizes a shared `configureJacoco` function with the `ApplicationAndroidComponentsExtension` for further JaCoCo setup.