build: Update buildscript and plugins configuration
The `buildscript` block has been added to the top-level `build.gradle.kts` file. This block configures repositories for Google and Maven Central, and adds a classpath dependency for `libs.google.oss.licenses.plugin`.

The `plugins` block has been moved to the end of the file and updated with several changes:
- `libs.plugins.kotlin.compose` updated to `libs.plugins.compose`
- `libs.plugins.kotlin.ksp` updated to `libs.plugins.ksp`
- `libs.plugins.google.services` updated to `libs.plugins.gms`
- Added new plugins:
    - `libs.plugins.kotlin.jvm`
    - `libs.plugins.dependency.guard`
    - `libs.plugins.firebase.crashlytics`
    - `libs.plugins.kotlin.serialization`
    - `libs.plugins.firebase.perf`
    - `libs.plugins.secrets`